实现动态获取（根据配置文件）并使用序列化器（jdk，json，kryo，hessian）

+++ common/src/main/java/com/scut/common/service/UserService.java
@@ -18,7 +18,7 @@ public interface UserService {
     /**
      * 新方法-获取数字
      */
-    default short getNumber(){
+    default long getNumber(){
         return 1;
     }
 }
